+++ packages/ripple/src/ripple-handlers.ts
@@ -18,8 +18,8 @@ limitations under the License.
 import {RippleInterface} from '@material/mwc-base/utils.js';
 
 export interface RippleAPI extends RippleInterface {
-  handleMouseEnter: () => void;
-  handleMouseLeave: () => void;
+  startHover: () => void;
+  endHover: () => void;
 }
 
 /**
@@ -38,7 +38,7 @@ export interface RippleAPI extends RippleInterface {
  *
  *   render() {
  *     return html`
- *       <div @mousedown=${this.rippleHandlers.activate}></div>
+ *       <div @mousedown=${this.rippleHandlers.startPress}></div>
  *       ${this.renderRipple ? html`<mwc-ripple></mwc-ripple>` : ''}
  *     `;
  *   }
@@ -46,44 +46,44 @@ export interface RippleAPI extends RippleInterface {
  * ```
  */
 export class RippleHandlers implements RippleAPI {
-  activate: (ev?: Event) => void;
-  deactivate: () => void;
-  handleFocus: () => void;
-  handleBlur: () => void;
-  handleMouseEnter: () => void;
-  handleMouseLeave: () => void;
+  startPress: (ev?: Event) => void;
+  endPress: () => void;
+  startFocus: () => void;
+  endFocus: () => void;
+  startHover: () => void;
+  endHover: () => void;
 
   constructor(
       /** Function that returns a `mwc-ripple` */
       rippleFn: () => Promise<RippleAPI|null>) {
-    this.activate = (ev?: Event) => {
+    this.startPress = (ev?: Event) => {
       rippleFn().then((r) => {
-        r && r.activate(ev);
+        r && r.startPress(ev);
       });
     };
-    this.deactivate = () => {
+    this.endPress = () => {
       rippleFn().then((r) => {
-        r && r.deactivate();
+        r && r.endPress();
       });
     };
-    this.handleFocus = () => {
+    this.startFocus = () => {
       rippleFn().then((r) => {
-        r && r.handleFocus();
+        r && r.startFocus();
       });
     };
-    this.handleBlur = () => {
+    this.endFocus = () => {
       rippleFn().then((r) => {
-        r && r.handleBlur();
+        r && r.endFocus();
       });
     };
-    this.handleMouseEnter = () => {
+    this.startHover = () => {
       rippleFn().then((r) => {
-        r && r.handleMouseEnter();
+        r && r.startHover();
       });
     };
-    this.handleMouseLeave = () => {
+    this.endHover = () => {
       rippleFn().then((r) => {
-        r && r.handleMouseLeave();
+        r && r.endHover();
       });
     };
   }